A0-2985: Refactor input address resolve domains
